requirements.txt
flask==3.0.0
replicate==0.25.1
python-dotenv==1.0.0
Pillow>=10.0.0
